libsemanage: add mls and target-platform options to semanage.conf
Whether or not to build an mls policy used to be determined by whether
or not the base module was mls. Base modules are no longer a thing, and
no modules are specific to either mls or non-mls, so that information is
now stored in semanage.conf.

Similarly, the target platform used to come from the base module. This
is going away, so that information must from from libsemanage or command
line tools.